# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MindCitas API",
    "version": "1.0.0",
    "description": "API REST de la plataforma de bienestar emocional MindCitas para estudiantes universitarios colombianos"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Desarrollo local"
    },
    {
      "url": "https://mindcitas-api.onrender.com/api",
      "description": "Produccion (Render)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "academicProgram": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "student",
              "professional",
              "admin"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "HabitSummary": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "habitType": {
            "type": "string"
          },
          "currentStreak": {
            "type": "integer"
          },
          "bestStreak": {
            "type": "integer"
          },
          "completedToday": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "security": [],
  "paths": {
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Lista usuarios registrados (paginado)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Filtro por email o nombre"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "No es administrador"
          }
        }
      }
    },
    "/admin/users/{id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Activa o desactiva un usuario",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "isActive"
                ],
                "properties": {
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado"
          },
          "400": {
            "description": "No puedes desactivar tu propia cuenta"
          },
          "403": {
            "description": "No es administrador"
          }
        }
      }
    },
    "/admin/stats": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Metricas generales de la plataforma",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Estadisticas de la plataforma",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalUsers": {
                          "type": "integer"
                        },
                        "activeUsers": {
                          "type": "integer"
                        },
                        "totalSessions": {
                          "type": "integer"
                        },
                        "sessionsLast7Days": {
                          "type": "integer"
                        },
                        "totalEmotionalEntries": {
                          "type": "integer"
                        },
                        "entriesLast7Days": {
                          "type": "integer"
                        },
                        "totalHabits": {
                          "type": "integer"
                        },
                        "completedTodayCount": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "No es administrador"
          }
        }
      }
    },
...
    "/habits": {
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Crea un nuevo habito para el usuario",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "habitType"
                ],
                "properties": {
                  "habitType": {
                    "type": "string",
                    "enum": [
                      "meditation",
                      "exercise",
                      "reading",
                      "hydration",
                      "sleep",
                      "custom"
                    ]
                  },
                  "name": {
                    "type": "string",
                    "example": "Meditacion matutina"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Habito creado"
          },
          "400": {
            "description": "Error de validacion"
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
    "/habits/mine": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "Lista los habitos activos del usuario",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de habitos con estado de racha",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HabitSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
    "/habits/{id}/streak": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "Detalle de racha de un habito",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de racha",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "currentStreak": {
                          "type": "integer"
                        },
                        "bestStreak": {
                          "type": "integer"
                        },
                        "completedDates": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "date"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "No tienes acceso a este habito"
          },
          "404": {
            "description": "Habito no encontrado"
          }
        }
      }
    },
    "/habits/{id}/check": {
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Marca el habito como completado hoy",
        "description": "Idempotente. Si ya fue marcado hoy, retorna el mismo estado.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado del check",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "currentStreak": {
                          "type": "integer"
                        },
                        "bestStreak": {
                          "type": "integer"
                        },
                        "awardedBadge": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "No tienes acceso a este habito"
          }
        }
      }
    },
    "/habits/{id}": {
      "delete": {
        "tags": [
          "Habits"
        ],
        "summary": "Desactiva un habito (soft delete)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Habito desactivado"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "No tienes acceso a este habito"
          }
        }
      }
    },
...
  }
}